Fix MalformedTitleException on change file name form
When you enter an invalid title you are asked to fix it. The form
needs to show the invalid title. To do this it tries to comnvert it
to a Title object first, which can't work.

I think this is a regression but didn't bother checking when it was
introduced.

I also made sure you can rename the file to "0". That's a special
case to consider because it equals false in PHP.

Bug: T314065
Change-Id: Ia594941621b1651f71c0cf8381d68538d9f7d388