Ignore only the first prefix from e.g. "Categorie:File:…"
This "strip prefix" method is only used for values from API
responses. These are guaranteed to have the namespace prefix. But
it's localized. That's why this code does a simple explode and
doesn't try to create an actual Title object. That wouldn't work.
But it shouldn't strip later parts from category and template names
after the namespace. There are actually categories like
[[Category:User:en]] and such on many wikis.

See https://www.mediawiki.org/wiki/Special:Diff/5472198 for the
edit request that triggered this.

Change-Id: Id804202a3a7bbae12b03872d1e6fdaafc0f2c3cd