Consistent terminology and slightly more helpful error messages
Notably: "transfer" is a word that doesn't appear anywhere else.
What we do from the perspective of this extention is an import.

Two more error messages are made more actionable by adding the most
likely interpretation as an example.

This patch also fixes a minor PHP8 compatibility issue. fopen() can
apparently throw an Error, which is not an Exception, but both are
Throwables.

Change-Id: Iab32a5bb3471151f11c76468c3dc05c481b2a73c