Inline/compact overly complex pieces of code
The check for an $allowedDomain needs a bit of an explanation. Before,
the strict `$host === $allowedDomain` was only done when the pattern
does not start with a dot. Turns out this was irrelevant. When the
pattern starts with a dot, e.g. ".wikipedia.org", anything that ends
with ….wikipedia.org is allowed, including the string that is
identical to ".wikipedia.org". That's why we can do the identity
check first.

Depends-On: Icf0d13a79a4ce5a72ac8459fde88d7ec04d073d2
Change-Id: If80eabd52021346a8184af5823e23d3991c01d10