Avoid DB access in tests when possible
getTestUser() is not a conveniency method for getting a mock User. The
TestUser is created in the test database, and therefore the test needs a
DB connection for it to work. None of these tests actually need a real
user, so just use a mock instead.

SpecialImportFileIntegrationTest probably also doesn't need a real user
and it may just use Authority, but that seems not to work, so add the
test to the Database group with a TODO instead.

The depends-on is needed for the typehints it adds to some User and
WebRequest methods. Otherwise we would need to explicitly mock their
return values in several tests, or other pieces of code will fail if
they call one of those methods and the return value is not of the
expected type.

Depends-On: I80723b886b2b5a5d75cbb73571e1b19ea4a09af5
Change-Id: If56144374d76c3b79549870d6b282b47fe59a686