Change ImportPlanValidatorTest to use qqx language code
The fact that LocalizedImportException hard-coded English in a way
that made it impossible to use it with the qqx dummy language code
always bothered me. I think I found a good way now. The idea is that
the internal Exception::getMessage() string (meant to appear in e.g.
logs) should not be localized but either be canonical English or qqx.

Only ILocalizedException::getMessageObject() is localized.

Note this was already the case. This patch:
* Lifts the hard-coded "en" and allows both "qqx" and "en" now.
* Now uses plain() without any transformations.

Change-Id: Ia4bb7cf033a3ce924fad182df3ff771b24540833