NullRevisionCreatorTest: use valid title for test
Recent versions of MW throw an InvalidArgumentException on the
testCreateForLinkTargetSuccess test: "InvalidArgumentException:
WikiPage constructed on a Title that cannot exist as a page:".
There were two reasons for that:

1. We were using __METHOD__ in the title for the test case, and
that contains a double colon, which makes for an invalid title.  Oops.

2. NullRevisionCreator::publishLogEntry() calls Revision::getPage() to
set the title target for the log entry, and we were only mocking
Revision::getPageAsLinkTarget(), not ::getPage().  As a result
we were ending up with an empty Title.

Added a proper mock for Revision::getPage(), and went ahead and added
one for Revision::getTimestamp() as well, since that method is also
used by NullRevisionCreator::publishLogEntry().

Change-Id: I62f0b590d923d98b32cff5ed97f2d113e22748f7