Mock RestrictionStore in all tests
I can apparently be expensive to call this service, and we just
don't need to in any of these tests. This is inspired by I1de43b6.

Yes, this means there is no test case that covers the code where
we use RestrictionStore. The relevant piece of code was added in
I040610a with no tests. We might want to continue working on this
in the future. For now it should be more obvious that this is not
tested.

Change-Id: Ia109caf9d07815ebba37815198726a13b80ce520