Stop using core's ILocalizedException interface
As far as I can tell there was never really a point in using the core
interface other than convenience, or just having a pattern to follow.
Turns out that pattern was not so great, see T287405.

All the exceptions this code cares about are part of this codebase and
direct subclasses of LocalizedImportException. Core does have a few
more that implement the ILocalizedException interface, but these are
not relevant here.

Bug: T287405
Change-Id: I80d561e0249925a3caf0934a65d5c71526039266