Reduce confusion around duplicate InvalidArgumentException
There are two. The places touched in this patch are supposed to use
the normal one. The idea of this patch is to make this much easier
to see in the code.

But there is also \FileImporter\Exceptions\InvalidArgumentException,
a subclass of \FileImporter\Exceptions\ImportException, meant to be
catched and presented to the user. Introduced in 2017 via I66ce495.
We should probably rename it. But this is for a later patch.

Change-Id: I8c2183d6e4ae58ec17155258641b8869d668b0b4