Codex: Load the diff dynamically using the api
This patch makes use of the api action "compare" to display
the diff between the original and the current file description:
https://www.mediawiki.org/wiki/API:Compare

It makes some of the php code obsolete, that was creating the diff view.
We passed the initialWikitext from php to codex to make it available
and renamed title and edit summary accordingly.

To display the output of the api, we added a table to the template
(as described in the api docu).

Bug: T347457
Change-Id: Iccc638ec6730953639ca6bdfc24deca0ef593ec8