Codex: Display warning about file title in form instead of extra page
On change the title field is checking if the title is okay
(e.g. no lower case letter as the first letter).
If not it changes the title directly and displays a warning.
Unsetting the status on input makes sure the message can
be updated when the user enters a new title.

Bug: T348741
Change-Id: I632ea7da982c9814392527646a97153c837ceb7f