Codex: Dynamic form submit for FileImporter
We are using ajax to pass the form data to the special page. The
special page then does the import and returns pure JSON as result.

The frontend retrieves the JSON response from the page and shows
a dynamic result. For now we ignore the in depth error handling.

The errors that we don't handled yet won't return valid JSON and
so they will always trigger the error case in the frontend.

TODO (for separate patches):
- remove parts of the preview when successful?

Bug: T347458
Change-Id: Ib76d133e339b9e064a5e8fd46e68e6573b2800ca