Show readable message from HttpRequestException errors
MWHttpRequest requests return a status object than includes
several methods used to present the issue in a user readable way.

This method is also used in core to present errors triggerd by an
upload. See /includes/specials/SpecialUpload.php#504

Let's use this method to show the error in the interface.

Bug: T253008
Change-Id: I59f9541536baebb757b6e15602953fa4e2a3ad8b