Avoid growing the button when the field grows
Before the patch the button grew in height when the field grew. This
is a side effect from the flex layout used on these elements. The
field and the button are next to each other in a row. Then we use
justify-content: space-between; so that they are spread out and the
button right aligned.

But the latter also will stretch the button if the field grows. So
wrapping the button in an additional div will let that div grow
and the button will just stay as it is.

Bug: T348741
Change-Id: Ib28375695144d131058a195263ba2e4b7353b95a