Restore default values for missing content fields
This got lost in patch If1c4d99. A closer look at the code in
ApiQueryRevisionsBase that populates this part of the API response
shows that the fields can be missing for various reasons. For
example:
* Several elements can be "hidden" (partially deleted by an
  administrator). This is already taken care of in
  ApiDetailRetriever.
* A slot can be marked as being entirely "missing" in unknown error
  situations.
* There is also "textmissing" when a slot is there, but the content
  can't be deserialized for an unknown reason.
* "badcontentformat" is another unexpected error situation where the
  content is there, but in an unexpected format.

Such errors typically only happen with very, very old revisions.

FileImporter's general philosophy is to fail when one of the binary
files is not accessible, but be fine with partly missing text
information. It's much better to still import the remaining
information, most notably the file and text revision histories with
timestamps and links to the user accounts. The alternative is that
users switch to the old process where all this information is lost.

Bug: T352984
Change-Id: I97ba475c98711a27b9ec356dd40dfdd220d77051