Drop a bunch of unused test setup and mocks
Some of this was used at some point, but isn't any more because the
corresponding code changed. The use of createNoOpMock in Id57631c
helped a lot and made all this much more obvious.

Also replace a whole bunch of deprecated class aliases. I checked
and these are all aliases from 1.41 and 1.42. The minimum version
in extension.json is already 1.42 since If23d9e0.

Change-Id: I93d6c32abfca995c1140aac259323d1d528412ed