Add language-level return type `string` to all getHtml methods
This is split from I90f3483. It should be rather straightforward,
so I hope. The only thing to consider is that the old code made it
possible to return e.g. an OOUI widget instance or something else
with a magic __toString() method. But this code here never did that.
We want it to return only strings.

Change-Id: Ic727c88742f8b5ac47ace6efd644344246f1a452