Faster wikitext template parsing with native strcspn()
Same as in If7f28a0. The idea is that the loop cares only for very
few special characters. Everything else is completely ignored and
doesn't do anything other than incrementing $i. While this works
totally fine with the for loop in PHP it can be slow when there are
thousands of irrelevant characters.

strcspn() is a low-level C function made exactly for this use case.

This probably doesn't have much of an effect in real-world scenarios.
Templates on file pages are usually short. Still worth it in my
opinion.

This code is heavily covered with tests. There is no danger of this
doing anything unexpected.

Change-Id: Id1a26131d1b17966b2ee1500c9baaae10e59fc80