Minimize not needed setup in FileImporterUploadBaseTest
There is no point in adding text to the images. Nobody can ever see
this. They can also be much smaller to save memory and setup time.
The only thing I leave behind is a single color so it's easier to see
is on a black or white background. Useful for debugging.

Change-Id: Ic06f71514ce1fbe09dbff154da9698b34fc2db18