Avoid string offset error when parsing unclosed templates
This issue is caused by the optimization added via Id1a2613. The
error does have zero impact on users. It is just a notics in our
logs. The code continues to work just fine.

We could add an extra if() to check if $i is already at the end of
the string. The null check suggested in patchset 1 does have the
same effect. The loop will do an extra, meaningles iteration because
of this, but not do anything. I think that's fine. It happens only
in this rare edge-case.

Bug: T376878
Change-Id: I83bf335388af98897c14679e51275169703f45ff