Fix incomplete "cannot parse URL $1" message
The message shows the parameter placeholder $1 unresolved because the
value is missing. This is apparently an issue since 2017 when the
parameter was added to the message, see I1610f83. The patch forgot to
update the existing place that already used the message.

Users probably never see this. I have seen it in a PHPUnit test run.

Change-Id: Ie86b2217a81a0daae405196c7c0db03530ee9b94