Replace confusing array_diff in ImporterTest
The first array contains the elements we expect. When the second
array contains them as well (and possibly more) they will all be
removed from the first array and the result will be empty.

This is somewhat convoluted for a "contains" assertion. PHPUnit
gives us better terminology to do the same.

Change-Id: I2be4102b6cee87aca29c9b1aadaa1d6e055a33ea