Change not-needed protected properties to private
These properties are marked as "protected" ever since the classes
have been created way back in 2017 via Ib38a252 and later via
I8df01d0. "Protected" means "public to subclasses" and should be
avoided for the same reasons "public" should be avoided when it's
not really needed. I think this here was just a mistake or out
habbit.

Change-Id: Ia69a9516f4470720a02ef4df20f546bf8770db28