Fix docs of nullable strings in CommonsHelperConfigRetriever
These two properties are guaranteed to be set after
CommonsHelperConfigRetriever::retrieveConfiguration() returned
true. The two getters should (and are) only be called after
retrieveConfiguration() was successful.

The extra initialization with null is not needed because that's
the default anyway. And as said, nobody is allowed to access these
anyway when they are null.

Change-Id: I71d60d745c0bc853d4ed71c44a7d7dcbee0daaf7