Stop marking all imported revisions as minor
This code is problematic since we switched to formatversion=2 via
If1c4d99. That was about 2 years ago. All imported revisions are
marked as minor since then. This happens because the "minor"
element is now always present as either "true" or "false". With
formatversion=1 the element was missing.

This might also be related to other, otherwise unrelated changes in
how core stores and handles the "minor" flag.

Bug: T403875
Change-Id: I4a643bb3c6d7fee3f8aa1008e937cb2524c34e30