[selenium] Drop selenium browser test
We currently have only one e2e test in this codebase that's testing
if the help popup is displayed and vanishes correctly.

E2e tests take up a lot of time in the CI runs and we do not consider
the aspect tested here as very essential or the implementation as
very fragile.  That's why we decided on dropping it.

Change-Id: I1ff534449f14c46dc3478ee682f6b84a5977f1c0